Update timer utility module
